Fix: Ajuste para buscar informações para graficos

+++ src/main/java/com/bazar/api/domain/movimentacao/dto/DadosGrafico.java
@@ -1,8 +1,10 @@
 package com.bazar.api.domain.movimentacao.dto;
 
+import java.math.BigDecimal;
+
 public record DadosGrafico(
 
         String periodo,
-        Long total
+        BigDecimal total
 ) {
 }
